requirements.txt
nltk==3.8.1
matplotlib==3.8.3
seaborn==0.13.2
pandas==2.2.1
wordcloud==1.9.3
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.14.0
tqdm==4.67.1
python-dotenv==1.0.1
